homework-5/app/requirements.txt
django
psycopg2-binary
python-dotenv
Pillow
djangorestframework